pyproject.toml
[tool.poetry]
name = "thumbhash-python"
version = "0.1.1"
description = "An experimental port of ThumbHash to Python"
authors = ["Lukas Winkler <git@lw1.at>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^9.4.0"
numpy = "^1.24.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
